created general folder to have generic response class for response objects to inherit from
