+++ backend/src/main/java/tech/geocodeapp/geocode/user/response/GetCurrentCollectableResponse.java
@@ -6,6 +6,7 @@ import tech.geocodeapp.geocode.collectable.model.Collectable;
 import io.swagger.v3.oas.annotations.media.Schema;
 import org.springframework.validation.annotation.Validated;
 import tech.geocodeapp.geocode.collectable.model.Collectable;
+import tech.geocodeapp.geocode.general.Response;
 
 import javax.validation.Valid;
 import javax.validation.constraints.*;
@@ -17,61 +18,13 @@ import javax.validation.constraints.*;
 @javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2021-06-24T10:12:19.520Z[GMT]")
 
 
-public class GetCurrentCollectableResponse   {
-  @JsonProperty("success")
-  private Boolean success = null;
-
-  @JsonProperty("message")
-  private String message = null;
-
+public class GetCurrentCollectableResponse extends Response {
   @JsonProperty("Collectable")
   private Collectable collectable = null;
 
   public GetCurrentCollectableResponse(boolean success, String message, Collectable collectable) {
-    this.success = success;
-    this.message = message;
+    super(success, message);
     this.collectable = collectable;
-  }
-
-
-  public GetCurrentCollectableResponse success(Boolean success) {
-    this.success = success;
-    return this;
-  }
-
-  /**
-   * Get success
-   * @return success
-   **/
-  @Schema(required = true, description = "")
-      @NotNull
-
-    public Boolean isSuccess() {
-    return success;
-  }
-
-  public void setSuccess(Boolean success) {
-    this.success = success;
-  }
-
-  public GetCurrentCollectableResponse message(String message) {
-    this.message = message;
-    return this;
-  }
-
-  /**
-   * Get message
-   * @return message
-   **/
-  @Schema(example = "The user's Collectable was successfully returned", required = true, description = "")
-      @NotNull
-
-    public String getMessage() {
-    return message;
-  }
-
-  public void setMessage(String message) {
-    this.message = message;
   }
 
   public GetCurrentCollectableResponse collectable(Collectable collectable) {
@@ -93,38 +46,6 @@ public class GetCurrentCollectableResponse   {
 
   public void setCollectable(Collectable collectable) {
     this.collectable = collectable;
-  }
-
-
-  @Override
-  public boolean equals(java.lang.Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-    GetCurrentCollectableResponse getCurrentCollectableResponse = (GetCurrentCollectableResponse) o;
-    return Objects.equals(this.success, getCurrentCollectableResponse.success) &&
-        Objects.equals(this.message, getCurrentCollectableResponse.message) &&
-        Objects.equals(this.collectable, getCurrentCollectableResponse.collectable);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(success, message, collectable);
-  }
-
-  @Override
-  public String toString() {
-    StringBuilder sb = new StringBuilder();
-    sb.append("class GetCurrentCollectableResponse {\n");
-    
-    sb.append("    success: ").append(toIndentedString(success)).append("\n");
-    sb.append("    message: ").append(toIndentedString(message)).append("\n");
-    sb.append("    collectable: ").append(toIndentedString(collectable)).append("\n");
-    sb.append("}");
-    return sb.toString();
   }
 
   /**
