+++ backend/src/main/java/tech/geocodeapp/geocode/user/response/SwapCollectableResponse.java
@@ -2,17 +2,14 @@ package tech.geocodeapp.geocode.user.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import tech.geocodeapp.geocode.collectable.model.Collectable;
+import tech.geocodeapp.geocode.general.Response;
 
-public class SwapCollectableResponse {
-    @JsonProperty("success")
-    private Boolean success = null;
-
-    @JsonProperty("message")
-    private String message = null;
+public class SwapCollectableResponse extends Response {
 
     /**
      * The found collectable with the given collectableID
      */
+    @JsonProperty("collectable")
     private Collectable collectable;
 
     /**
@@ -25,11 +22,10 @@ public class SwapCollectableResponse {
     /**
      * Overloaded Constructor
      *
-     * @param collectable The collectable from the specified collectable
+     * @param collectable The collectable that is swapped out of the GeoCode
      */
     public SwapCollectableResponse( boolean success, String message, Collectable collectable ) {
-        this.success = success;
-        this.message = message;
+        super(success, message);
         this.collectable = collectable;
     }
 
@@ -51,21 +47,5 @@ public class SwapCollectableResponse {
     public void setCollectable( Collectable collectable ) {
 
         this.collectable = collectable;
-    }
-
-    public Boolean getSuccess() {
-        return success;
-    }
-
-    public void setSuccess(Boolean success) {
-        this.success = success;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
     }
 }
